Refactor Home
